MemberReport moved to MemberReport.cs
